fixing review utc date and added Person score from all roles

+++ Application/Features/People/PersonDto.cs
@@ -12,7 +12,7 @@ public record PersonDto(
     string Image,
     DateTime BirthDate,
     string Biography,
-    double Score,
+    decimal Score,
     int NumberOfReviews,
     IEnumerable<PersonRolesDto> Roles);
